projects/quotecard/requirements.txt
# QuoteCard Dependencies
google-genai>=0.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
